+++ lib/core/visual_constants.dart
@@ -1,3 +1,5 @@
+
+/// Formats values in MB or GB
 String formatSize(double sizeInBytes) {
   const oneMB = 1024 * 1024;
   const oneGB = 1024 * oneMB;